+++ Configuration/TCA/Overrides/tt_content.php
@@ -1,4 +1,7 @@
 <?php
+
+use TYPO3\CMS\Core\Information\Typo3Version;
+use TYPO3\CMS\Core\Utility\GeneralUtility;
 // all use statements must come first
 //use TYPO3\CMS\Core\Utility\ExtensionManagementUtility;
 
@@ -30,11 +33,20 @@ defined('TYPO3') or die();
         'Feuserregistration: Verify a registration.'
     );
 
+    $xmlFile = 'FILE:EXT:' . $extensionName . '/Configuration/Flexforms/Verify.xml';
+
+    $versionInformation = GeneralUtility::makeInstance(Typo3Version::class);
+
+    if (version_compare($versionInformation->getVersion(), '12.0', '<')) {
+        $pluginSignature = $extensionName . '_verify';
+        $xmlFile = 'FILE:EXT:' . $extensionName . '/Configuration/Flexforms/VerifyV11.xml';
+    }
+
     $GLOBALS['TCA']['tt_content']['types']['list']['subtypes_addlist'][$pluginSignature] = 'pi_flexform';
 
     \TYPO3\CMS\Core\Utility\ExtensionManagementUtility::addPiFlexFormValue(
         $pluginSignature,
-        'FILE:EXT:' . $extensionName . '/Configuration/Flexforms/Verify.xml'
+        $xmlFile
     );
      /******************************************************************
      * FRONTEND PLUGINS - END